1_5_2015
Commit Notes

Changes to CPLD_INIT

Logic level changed to '1' on the Octal buffer enable used in PowerController.vhd and
FlashInit.vhd. These are properly inverted at the top levels of _init and non_init
CPLD projects.

Changes to CPLD

fpga_turnon reduced to fpga_running and forced_start_in for the time being.
It was found that the real time clock is outputting a clock on its interrupt
line in its default state.
fpga_running has changed to an or of fpga_powered and conf_done and init_done.
It was found that conf_done fpga line goes high as the FPGA receives all data (minus 2 clock cycles)
and init_done fpga line (if programmed into FPGA stream and project) goes high after the
FPGA has completely entered user mode. When all three are high we can put the PFL loader
back into reset state.

nconfig_out was changed to handoff to PFL loader on fpga_running from fpga_activated to hand over control
to the PFL loader sooner in the process. However I believe this was not what fixed
the FPGA programming. In the associated logic analyzer runs, nconfig was never
seen to go low as is indicated by documentation. Vol 1 Chp 13 Cyclone V handbook.

The reason the FPGA was not programming was the reset of the PFL was occurring
too fast as well as init_done not being properly configured into the FPGA project.
init_done is a signal that the fpga_running uses, which in turn will put the PFL
loader back into a reset state.  init_done tells us we have successfully entered
user mode on the FPGA.

An associated logic analyzer run of all the signals associated with FPGA programming
was saved to chris's debug diary. It helps illuminate how the FPGA status lines
react after power up and in relation to the CPLD.

Changes to FPGA

Creation of base FPGA project file along with DevBoard_PowerMonitorFPGA_TopLevel.vhd
mapping into collar.vhd.

Edit to the _IO.qsf file (This is the .qsf that we import, not the one
maintained by Quartus.)

In future projects it may just be easier to write the bank voltages and
I/O standards into the qsf file rather than fight with Pin Planner, which just
writes the same things in anyway.

In the current state INIT_DONE must be enabled in the FPGA project.
The CPLD uses this INIT_DONE line to put the PFL back into a reset state at
the appropriate time. The current project has init_done enabled.

The current project has the selectable GPIO bank set to 2.5 volts. This
corresponds to the jumper settings on the board. The jumpers and pin planner
bank voltage need to be changed if the voltage is changed.

The FPGA top level has a counter output to selectable GPIO
which was tested first successfully on 1_2_2015.